% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bali}
\alias{bali}
\title{Bali terrorist network}
\format{
A \code{statnet}'s network class object.
data(bali)
}
\description{
The network corresponds to the contacts between the 17 terrorists
who carried out the bombing in Bali, Indonesia in 2002.
The network is taken from Koschade (2006).
}
\references{
Koschade, S. (2006). A social network analysis of Jemaah Islamiyah: The applications to counter-terrorism and intelligence.
Studies in Conflict and Terrorism, 29, 559--575.
}
