% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_between_networks}
\alias{get_between_networks}
\title{Obtain the between-block networks defined by the block attribute.}
\usage{
get_between_networks(network, block)
}
\arguments{
\item{network}{a network object}

\item{block}{a vector of integers representing the block of each node}
}
\value{
a list of networks
}
\description{
Function to return a list of networks, each network representing the within-block network of a block.
}
\examples{
# Load an embedded network object.
data(toyNet)
get_within_networks(toyNet, toyNet \%v\% "block")
}
