% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{kapferer}
\alias{kapferer}
\title{Kapferer collaboration network}
\format{
A \code{statnet}'s network class object.
data(kapferer)
}
\description{
The network corresponds to collaborations between 39 workers in a tailor shop in Africa:
an undirected edge between workers \code{i} and \code{j} indicates that the workers collaborated.
The network is taken from Kapferer (1972).
}
\references{
Kapferer, B. (1972). Strategy and Transaction in an African Factory. Manchester University Press, Manchester, U.K.
}
