% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot.bigergm}
\alias{plot.bigergm}
\title{Plot the network with the found clusters}
\usage{
\method{plot}{bigergm}(x, ...)
}
\arguments{
\item{x}{The output of the bigergm function}

\item{...}{Additional arguments, to be passed to lower-level functions}
}
\description{
This function plots the network with the found clusters. The nodes are colored according to the found clusters.
Note that the function uses the \code{network} package for plotting the network and should therefore not be used for large networks with more than 1-2 K vertices
}
