% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reed}
\alias{reed}
\title{A network of friendships between students at Reed College.
The data was collected by Facebook and provided as part of Traud et al. (2012)}
\format{
A \code{statnet}'s network class object. It has three nodal features.
\describe{
\item{doorm}{anonymized dorm in which each node lives.}
\item{gender}{gender of each node.}
\item{high.school}{anonymized highschool to which each node went to.}
\item{year}{year of graduation of each node.}
...
}
}
\usage{
reed
}
\description{
A network of friendships between students at Reed College.
The data was collected by Facebook and provided as part of Traud et al. (2012)
}
\references{
Traud, Mucha, Porter (2012). Social Structure of Facebook Network.
Physica A: Statistical Mechanics and its Applications, 391, 4165-4180
}
\keyword{datasets}
